Selección de archivo de configuración de Nginx

+++ nginx/command.sh
@@ -3,17 +3,21 @@ set -e;
 
 rm -f /etc/nginx/conf.d/default.conf;
 
-if [ -d "$NGINX_AVAILABLE_SITES" ] ; then
-    ln -s "$NGINX_AVAILABLE_SITES"/"$NGINX_SSL_CONFIG_FILE" /etc/nginx/conf.d/default.conf;
+if [[ -n "$NGINX_CONFIG_FILE" \
+    && -f "$NGINX_SSL_CONFIG_DATA/andino.key" \
+    && -f "$NGINX_SSL_CONFIG_DATA/andino.crt" ]]; then
+    echo "Using '$NGINX_CONFIG_FILE' configuration"
+    ln -s "$NGINX_AVAILABLE_SITES/$NGINX_CONFIG_FILE" /etc/nginx/conf.d/default.conf;
     openssl dhparam 2048 -out $NGINX_SSL_CONFIG_DATA/andino_dhparam.pem
 else
-    if [[ -n "$NGINX_CONFIG_FILE" ]]; then
-        echo "Using '$NGINX_CONFIG_FILE' configuration"
-        ln -s "$NGINX_AVAILABLE_SITES/$NGINX_CONFIG_FILE" /etc/nginx/conf.d/default.conf;
-    else
-        echo "Using '$NGINX_DEFAULT_CONFIG_FILE' default configuration"
-        ln -s "$NGINX_AVAILABLE_SITES/$NGINX_DEFAULT_CONFIG_FILE" /etc/nginx/conf.d/default.conf;
-    fi
+    echo "Using '$NGINX_DEFAULT_CONFIG_FILE' default configuration"
+    ln -s "$NGINX_AVAILABLE_SITES/$NGINX_DEFAULT_CONFIG_FILE" /etc/nginx/conf.d/default.conf;
+fi
+
+# Configure extended cache if specified
+if [ "$NGINX_EXTENDED_CACHE" = 'yes' ] ; then
+    chmod -R u+rwx $NGINX_SCRIPTS
+    "$NGINX_SCRIPTS"/extend_nginx.sh ;
 fi
 
 # Configure max_size cache option
